backend/paddleocr-text-extraction/requirements.txt
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Cython==3.0.12
decorator==5.2.1
exceptiongroup==1.2.2
fastapi==0.115.12
fire==0.7.0
fonttools==4.57.0
google-auth==2.39.0
google-genai==1.13.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
lazy_loader==0.4
lmdb==1.6.2
lxml==5.4.0
networkx==3.4.2
numpy==2.2.5
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==25.0
paddleocr==2.10.0
paddlepaddle==3.0.0
pillow==11.2.1
protobuf==3.20.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.4
pydantic_core==2.33.2
python-docx==1.1.2
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.2
shapely==2.1.0
simsimd==6.2.1
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
stringzilla==3.12.5
termcolor==3.1.0
tifffile==2025.3.30
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
